feat: Add a `Token` struct to allow the CLI to differentiate between `Bearer` and `APIKey` tokens (#66)
This PR adds a `Token` struct to allow the CLI to differentiate between the auth methods.

The CLI will have a different download flow depending on if the auth is using a bearer token vs an API key. For the bearer token the `team_name` is configured using the `cloudquery switch` command. For the API key, the key is associated with a team name already.