feat: Add method to return token type (#75)
Whilst the token itself now contains the token type, it is useful to be able to determine the type of tokens the token client will return without fetching a token first. An example user of this functionality will be the usage client, which needs to determine if an API call is needed to fetch the team name when the authentication is using an API key.

Refs: https://github.com/cloudquery/cloudquery-issues/issues/893